go_router with state
