+++ lib/presentation/providers/providers.dart
@@ -1,4 +1,3 @@
-
 export 'movies/movies_providers.dart';
 export 'movies/movies_reposity_provider.dart';
 export 'movies/movies_slider_provider.dart';
@@ -8,3 +7,4 @@ export 'actors/actors_bymovie_provider.dart';
 export 'actors/actors_repository_provider.dart';
 
 export 'search/search_movie_provider.dart';
+export 'movies/initial_loading_provider.dart';
