+++ lib/presentation/views/views.dart
@@ -1,0 +1,2 @@
+export './tabs/favorites_view.dart';
+export './tabs/home_view.dart';
