+++ lib/presentation/widgets/movies/movies_horizontal.dart
@@ -95,7 +95,7 @@ class _Slide extends StatelessWidget {
                   }
                   return GestureDetector(
                     child: FadeIn(child: child),
-                    onTap: () => context.push('/movie/${movie.id}'),
+                    onTap: () => context.push('/home/0/movie/${movie.id}'),
                   );
                 },
               ),
