+++ lib/presentation/widgets/shared/custom_appbar.dart
@@ -37,7 +37,7 @@ class CustomAppbar extends ConsumerWidget {
                 )
               ).then((movie) {
                 if(movie == null) return;
-                context.push('/movie/${movie.id}');
+                context.push('/home/0/movie/${movie.id}');
               });
             }, icon: const Icon(Icons.search))
           ]),
