+++ lib/presentation/widgets/shared/custom_bottom_navigation.dart
@@ -1,14 +1,19 @@
 
 
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigation extends StatelessWidget {
-  const CustomBottomNavigation({super.key});
+  final int currentIndex;
+
+  const CustomBottomNavigation({super.key, required this.currentIndex});
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
       elevation: 0,
+      currentIndex: currentIndex,
+      onTap: (index) => context.go('/home/$index'),
       items: const [
         BottomNavigationBarItem(
           icon: Icon(Icons.home_max),
