+++ lib/presentation/widgets/widgets.dart
@@ -1,7 +1,7 @@
-
 export 'shared/custom_appbar.dart';
 export 'shared/custom_bottom_navigation.dart';
 
 
-
+export './shared/loader.dart';
 export 'movies/movies_slider.dart';
+export './movies/movies_horizontal.dart';
